Update: shell config
Since I am using both zsh and bash consistently, I decided to make a
separate `common` file and source it from both zshrc and bashrc.

Another idea for improvement is to automatically source files matching
the hostname (if they exist), that would end the trouble of
machine-specific variables/aliases/configs. The same could be done to OS
specific configs, although the way it is implemented right now is fine.